requirements.txt
beautifulsoup4==4.5.1
click==6.7
cookies==2.2.1
cryptography==1.5
Flask==0.11.1
Flask-Cors==2.1.2
geopy==1.11.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
oauthlib==2.0.2
python-dotenv==0.6.4
python-forecastio==1.3.5
requests==2.13.0
requests-oauthlib==0.8.0
responses==0.5.1
six==1.10.0
TwitterSearch (1.0.1)